moved FeatureUnion to transformations.compose
